requirements.txt
# Core AI & API
anthropic>=0.20.0
openai>=1.12.0

# RAG Framework
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-faiss>=0.1.0

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.26.0

# Embeddings
sentence-transformers>=2.3.0
torch>=2.1.0
transformers>=4.37.0

# Document Processing
PyMuPDF>=1.23.0
unstructured[pdf]>=0.12.0
python-docx>=1.1.0
pdfplumber>=0.10.0

# Translation
googletrans==4.0.0rc1
# IndicTrans2 (install separately from github)

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Bot Interfaces
python-telegram-bot>=20.7
# whatsapp-api-client-python>=1.0.0  # Uncomment for WhatsApp

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
aiofiles>=23.2.0
python-jose>=3.3.0

# Data
pandas>=2.1.0
scikit-learn>=1.4.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Dev/Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.0.0
